feat(query): Improve explain test compatibility from 57% to ~65%
Fix execute explain metrics (planExecutions/sizeOfResult), top-level
aggregate decomposition, average node chain explain, inline array
aggregate sources, and execute metrics u64 casting across all plan nodes.

- Fix planExecutions to count top-level query operations (N+1) matching
  Go's operationNode Next() semantics
- Fix sizeOfResult to count number of query results, not documents
- Decompose top-level averageNode into sumNode + countNode + averageNode
- Wrap non-top-level averageNode explain in countNode → sumNode chain
- Add SumSourceMeta, MaxSourceMeta, MinSourceMeta for inline array agg
- Cast all explain metrics to u64 for Go compatibility
- Handle execute vs default explain for top-level aggregate nodes
- Strip aggregate iterations when unwrapping for topLevelNode
- Add stub FFI functions for p2p_add/remove/get_all_documents and
  truncate_collection
- Handle _commits system collection in execute explain path

+++ crates/query/src/plan/scan.rs
@@ -303,22 +303,21 @@ impl PlanNode for ScanNode {
     fn explain_execute_inner(&self) -> serde_json::Value {
         let mut obj = serde_json::Map::new();
 
-        // Go DefraDB execute format: iterations, docFetches, fieldFetches, indexFetches
         obj.insert(
             "iterations".to_string(),
-            serde_json::json!(self.exec_info.iterations),
+            serde_json::json!(self.exec_info.iterations as u64),
         );
         obj.insert(
             "docFetches".to_string(),
-            serde_json::json!(self.exec_info.docs_fetched),
+            serde_json::json!(self.exec_info.docs_fetched as u64),
         );
         obj.insert(
             "fieldFetches".to_string(),
-            serde_json::json!(self.exec_info.fields_fetched),
+            serde_json::json!(self.exec_info.fields_fetched as u64),
         );
         obj.insert(
             "indexFetches".to_string(),
-            serde_json::json!(self.exec_info.indexes_fetched),
+            serde_json::json!(self.exec_info.indexes_fetched as u64),
         );
 
         serde_json::Value::Object(obj)
